+++ previous-versions/common/RELEASES.md
@@ -19,3 +19,13 @@ Based on:
 - [python v0.35.0] previous-versions/common
 ### Releases
 - [PyPI v0.35.0] https://pypi.org/project/codat-common/0.35.0 - previous-versions/common
+
+## 2023-12-04 01:27:32
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Platform.yaml
+- Speakeasy CLI 1.125.0 (2.209.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v0.35.1] previous-versions/common
+### Releases
+- [PyPI v0.35.1] https://pypi.org/project/codat-common/0.35.1 - previous-versions/common